mkdir triton package and move triton files
